# ContentApiCore/wwwroot/json/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Open Data Hub Content Api",
    "description": "Open Data Hub Content Api based on .Net Core with PostgreSQL",
    "termsOfService": "https://opendatahub.readthedocs.io/en/latest/",
    "contact": {
      "name": "Open Data Hub Team",
      "url": "https://opendatahub.com/",
      "email": "help@opendatahub.com"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://api.tourism.testingmachine.eu"
    }
  ],
  "paths": {
...
    "/v1/Source": {
      "get": {
        "tags": [
          "Source"
        ],
        "summary": "GET Sources List",
        "parameters": [
          {
            "name": "pagenumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pagesize",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "language",
            "in": "query",
            "description": "Language field selector, displays data and fields available in the selected language (default:'null' all languages are displayed)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idlist",
            "in": "query",
            "description": "IDFilter (Separator ',' List of IDs, 'null' = No Filter), (default:'null')",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "typelist",
            "in": "query",
            "description": "Type Filter (Separator ',' Filter Sources by Type ('accommodation','event','odhactivitypoi','webcam','region','municipality','district','tourismassocation','weather','measuringpoint','weatherhistory','weatherforecast','accommodationroom','venue','article','eventshort','wineaward','tag','odhtag'), (default:'null')",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fields",
            "in": "query",
            "description": "Select fields to display, More fields are indicated by separator ',' example fields=Id,Active,Shortname (default:'null' all fields are displayed). <a href=\"https://github.com/noi-techpark/odh-docs/wiki/Common-parameters%2C-fields%2C-language%2C-searchfilter%2C-removenullvalues%2C-updatefrom#fields\" target=\"_blank\">Wiki fields</a>",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "searchfilter",
            "in": "query",
            "description": "String to search for, Title in all languages are searched, (default: null) <a href=\"https://github.com/noi-techpark/odh-docs/wiki/Common-parameters%2C-fields%2C-language%2C-searchfilter%2C-removenullvalues%2C-updatefrom#searchfilter\" target=\"_blank\">Wiki searchfilter</a>",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rawfilter",
            "in": "query",
            "description": "<a href=\"https://github.com/noi-techpark/odh-docs/wiki/Using-rawfilter-and-rawsort-on-the-Tourism-Api#rawfilter\" target=\"_blank\">Wiki rawfilter</a>",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rawsort",
            "in": "query",
            "description": "<a href=\"https://github.com/noi-techpark/odh-docs/wiki/Using-rawfilter-and-rawsort-on-the-Tourism-Api#rawsort\" target=\"_blank\">Wiki rawsort</a>",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "removenullvalues",
            "in": "query",
            "description": "Remove all Null values from json output. Useful for reducing json size. By default set to false. Documentation on <a href=\"https://github.com/noi-techpark/odh-docs/wiki/Common-parameters,-fields,-language,-searchfilter,-removenullvalues,-updatefrom#removenullvalues\" target=\"_blank\">Opendatahub Wiki</a>",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List created",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SourceLinked"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SourceLinked"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SourceLinked"
                  }
                }
              },
              "text/csv": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SourceLinked"
                  }
                }
              },
              "application/ld+json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SourceLinked"
                  }
                }
              },
              "application/ldjson": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SourceLinked"
                  }
                }
              },
              "application/rawdata": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SourceLinked"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Request Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/csv": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/ld+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/ldjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/rawdata": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      },
      "post": {
        "tags": [
          "Source"
        ],
        "summary": "POST Insert new Source",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/SourceLinked"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceLinked"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceLinked"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SourceLinked"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/v1/Source/{id}": {
      "get": {
        "tags": [
          "Source"
        ],
        "summary": "GET Source Single",
        "operationId": "SingleSource",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the Source",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "in": "query",
            "description": "Language field selector, displays data and fields available in the selected language (default:'null' all languages are displayed)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fields",
            "in": "query",
            "description": "Select fields to display, More fields are indicated by separator ',' example fields=Id,Active,Shortname (default:'null' all fields are displayed). <a href=\"https://github.com/noi-techpark/odh-docs/wiki/Common-parameters%2C-fields%2C-language%2C-searchfilter%2C-removenullvalues%2C-updatefrom#fields\" target=\"_blank\">Wiki fields</a>",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "localizationlanguage",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "removenullvalues",
            "in": "query",
            "description": "Remove all Null values from json output. Useful for reducing json size. By default set to false. Documentation on <a href=\"https://github.com/noi-techpark/odh-docs/wiki/Common-parameters,-fields,-language,-searchfilter,-removenullvalues,-updatefrom#removenullvalues\" target=\"_blank\">Opendatahub Wiki</a>",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Object created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SourceLinked"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceLinked"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceLinked"
                }
              },
              "text/csv": {
                "schema": {
                  "$ref": "#/components/schemas/SourceLinked"
                }
              },
              "application/ld+json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceLinked"
                }
              },
              "application/ldjson": {
                "schema": {
                  "$ref": "#/components/schemas/SourceLinked"
                }
              },
              "application/rawdata": {
                "schema": {
                  "$ref": "#/components/schemas/SourceLinked"
                }
              }
            }
          },
          "400": {
            "description": "Request Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/csv": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/ld+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/ldjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/rawdata": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      },
      "put": {
        "tags": [
          "Source"
        ],
        "summary": "PUT Modify existing Source",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Source Id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Source Object",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/SourceLinked"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceLinked"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceLinked"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SourceLinked"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Source"
        ],
        "summary": "DELETE Source by Id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Source Id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "LicenseInfo": {
        "type": "object",
        "properties": {
          "License": {
            "enum": [
              "CC0",
              "CC-BY",
              "Closed"
            ],
            "type": "string",
            "nullable": true
          },
          "LicenseHolder": {
            "type": "string",
            "nullable": true
          },
          "Author": {
            "type": "string",
            "nullable": true
          },
          "ClosedData": {
            "type": "boolean",
            "description": "readonly field",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "Metadata": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "string",
            "nullable": true
          },
          "Type": {
            "enum": [
              "accommodation",
              "accommodationroom",
              "event",
              "odhactivitypoi",
              "measuringpoint",
              "webcam",
              "article",
              "venue",
              "eventshort",
              "experiencearea",
              "region",
              "metaregion",
              "tourismassociation",
              "municipality",
              "district",
              "area",
              "wineaward",
              "skiarea",
              "skiregion",
              "odhtag",
              "publisher",
              "tag",
              "weatherhistory",
              "weather",
              "weatherdistrict",
              "weatherforecast",
              "weatherrealtime",
              "snowreport",
              "odhmetadata",
              "package",
              "ltsactivity",
              "ltspoi",
              "ltsgastronomy"
            ],
            "type": "string",
            "nullable": true
          },
          "LastUpdate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "Source": {
            "type": "string",
            "nullable": true
          },
          "Reduced": {
            "type": "boolean"
          },
          "UpdateInfo": {
            "$ref": "#/components/schemas/UpdateInfo"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "Type": {
            "type": "string",
            "nullable": true
          },
          "Title": {
            "type": "string",
            "nullable": true
          },
          "Status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "Detail": {
            "type": "string",
            "nullable": true
          },
          "Instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "SourceLinked": {
        "type": "object",
        "properties": {
          "_Meta": {
            "$ref": "#/components/schemas/Metadata"
          },
          "Self": {
            "type": "string",
            "description": "generated field",
            "nullable": true,
            "readOnly": true
          },
          "LicenseInfo": {
            "$ref": "#/components/schemas/LicenseInfo"
          },
          "Id": {
            "type": "string",
            "nullable": true
          },
          "Key": {
            "type": "string",
            "nullable": true
          },
          "Name": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "nullable": true
          },
          "Description": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "nullable": true
          },
          "FirstImport": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "LastChange": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "Url": {
            "type": "string",
            "nullable": true
          },
          "Interfaces": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Interfaces that are offered by the source",
            "nullable": true
          },
          "Types": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateInfo": {
        "type": "object",
        "properties": {
          "UpdatedBy": {
            "type": "string",
            "nullable": true
          },
          "UpdateSource": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "password": {
            "tokenUrl": "https://auth.opendatahub.testingmachine.eu/auth/realms/noi/protocol/openid-connect/token",
            "scopes": { }
          },
          "clientCredentials": {
            "tokenUrl": "https://auth.opendatahub.testingmachine.eu/auth/realms/noi/protocol/openid-connect/token",
            "scopes": { }
          }
        }
      }
    }
  },
...
}